Oakの実装 イベントの調整

+++ src/main/java/net/hennabatch/hennadungeon/effect/BuffEffect.java
@@ -4,7 +4,7 @@ import net.hennabatch.hennadungeon.entity.Status;
 
 public class BuffEffect extends StatusEffect{
 
-    public BuffEffect(int durationTime, Status.EnumStatus targetStatus, int val, Boolean isMagnification) {
+    public BuffEffect(int durationTime, Status.EnumStatus targetStatus, double val, Boolean isMagnification) {
         super(durationTime, targetStatus, val, isMagnification);
     }
